+++ drizzle.config.ts
@@ -1,4 +1,4 @@
-import "dotenv/config";
+// import "dotenv/config";
 import { defineConfig } from "drizzle-kit";
 
 import env from "./lib/env";
